Update with some remarks for Yogi
Yogi, the text under Node.js vs Browser, item 1: Full user-level system access., is unclear. I do not understand what we want to explain the students. They probably do not understand what 'runtime environment' is, and here we add other things they do not understand: 'unrestricted access to the network", "UI", "electron". 

Also, this sentence "Unlike the browser where Javascript runs in an isolated environment in which potentially unsafe software code can execute without affecting network resources or local applications. " does not make sense to me.

Under 'The DOM' we use "tree of nodes" but I think the students do not yet know what a node is.

Exercise 1: "Look for the files `index.html` and `script.js` in the folder scripts" - the index.html is not in the folder scripts so I changed this line.
Exercise 5: "Hint: Use `createTodo` function inside second parameter of addEventListener()" - I do not understand this