Changes & additions with remarks for Yogi, until Object built-in methods
I added a horizontal line to divide this section from Part 1. It is so much information that I think this will help when they have to scroll up and down.

I think it would be useful to tell them why the oops.js file is called like that. They do not know about OOP. If we do not discuss OOP in this project, then maybe change the name of this file to objectScript.js or something like that.

Do they know how to run code with the node command already? I cannot remember ;-). The previous part is run with LiveServer, I think.

The sentence "This is in contrast to objects instantiated from classes, which we'll look at later on." in Object Basics will have to be changed or taken out once we decide where we put classes.

In Dot notation this sentence: "The object name (in this case person) acts as the namespace" is not clear to me and I wonder if the students will understand it.

Last sentence in Dot notation: "Otherwise your methods will no longer work." I have added "Why do you think this is necessary?" because I think it would help them to think about the logic themselves and come up with the reason why it would not work.

Bracket notation: I think we should explain why sometimes you can only use bracket notation. It is added as a Note to Setting object members (last line) but I think it is not just a note, it is important enough to discuss it here because it can lead to a lot of confusion and mistakes.

Setting object members: I may have a dirty mind, but using the name 'Cratchit' must be a joke, no??

The section on "this" can be removed if we move OOP to the other project. We have to check in this entire project if we move OOP, because it comes back in several places.